pyproject.toml
[project]
name = "nagaagent"
version = "3.0.5"
description = "NagaAgent 3.0 - 智能对话助手，支持多MCP服务、流式语音交互、GRAG知识图谱记忆系统、RESTful API接口、控制台托盘功能"
authors = [
    {name = "柏斯阔落", email = "1781393885@qq.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # 核心依赖
    "mcp>=1.6.0",
    "openai>=1.76.0",
    "openai-agents>=0.0.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "aiohttp>=3.11.18",
    "pytz>=2024.1",
    "colorama>=0.4.6",
    "python-dateutil>=2.9.0.post0",
    # GRAG知识图谱相关依赖
    "py2neo>=2021.2.3",
    "pyvis>=0.3.2",
    "matplotlib>=3.10.0",
    # API服务器相关依赖
    "flask-cors>=6.0.1",
    "flask>=3.1.0",
    "gevent>=25.5.1",
    "edge-tts>=7.0.2",
    "fastapi>=0.115.0",
    # 网络通信相关依赖
    "uvicorn[standard]>=0.34.0",
    "librosa",
    "websockets>=14.0", # WebSocket通信支持
    # AI机器学习相关依赖
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "tqdm>=4.67.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "transformers>=4.51.3",
    # 数据处理相关依赖
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "griffe>=1.7.3",
    "anyio>=4.9.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "sse-starlette>=2.3.3",
    "starlette>=0.46.2",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.1",
    "idna>=3.10",
    "urllib3>=2.4.0",
    "typing-extensions>=4.13.2",
    "markdown>=3.8",
    # GUI界面相关依赖
    "pyqt5-qt5 ==5.15.2",
    "pyqt5==5.15.11",
    "playwright>=1.52.0",
    "greenlet>=3.2.1",
    "pyee>=13.0.0",
    "pygame>=2.6.0",
    "html2text>=2020.1.16",
    # 音频处理相关依赖
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pyaudio>=0.2.14",
    "emoji>=2.8.0",
    # 系统托盘相关依赖
    "pystray>=0.19.4",
    "pillow>=10.0.0",
    # 其他工具依赖
    "tiktoken",
    "bottleneck>=1.3.6",
    # 系统控制相关依赖
    "screen-brightness-control", # 屏幕亮度控制
    "pycaw", # 音频控制
    "comtypes", # pycaw的依赖
    # MCP工具相关依赖
    "jmcomic>=2.6.5", # 漫画下载工具
    "fastmcp>=2.11.2", # MCP工具框架
    # MQTT通信相关依赖
    "paho-mqtt>=2.0.0", # MQTT客户端库
    # 其他依赖
    "python-docx>=1.2.0",
    "mqtt-tool>=0.1.1",
    # Live2D相关依赖
    "live2d-py>=0.3.2",
    "pyopengl>=3.1.0",
    "pyopengl-accelerate>=3.1.0",
]

# 可选依赖组
[project.optional-dependencies]
audio = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pyaudio>=0.2.14",
]

system-control = [
    "screen-brightness-control",  # 屏幕亮度调节
    "pycaw",  # 系统音量调节（Windows）
    "comtypes",  # pycaw依赖
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["apiserver","mcpserver","mqtt_tool", "agent", "summer_memory", "thinking", "ui", "voice" ]

[tool.ruff]
line-length = 120 

